backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
spotipy
python-dotenv
spotdl
apscheduler
sse-starlette
aiosqlite
pydantic-settings
mutagen

# Testing
pytest
pytest-asyncio
httpx

# Resilience
tenacity

# Migrations
alembic

# Security
cryptography
